requirements.txt
boto3
requests
google-api-python-client
google-auth
python-dotenv